changing float to double
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@51011 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/main/java/org/geogebra/common/awt/GBasicStroke.java
@@ -12,12 +12,12 @@ public interface GBasicStroke {
 
 	int getEndCap();
 
-	float getMiterLimit();
+	double getMiterLimit();
 
 	int getLineJoin();
 
-	float getLineWidth();
+	double getLineWidth();
 
-	float[] getDashArray();
+	double[] getDashArray();
 
 }
