+++ desktop/src/main/java/org/geogebra/desktop/awt/GGradientPaintD.java
@@ -24,9 +24,10 @@ public class GGradientPaintD implements GGradientPaint {
 						copyg.getColor2().getAlpha()));
 	}
 
-	public GGradientPaintD(float x1, float y1, GColor color1, float x2,
-			float y2, GColor color2) {
-		impl = new GradientPaint(x1, y1, GColorD.getAwtColor(color1), x2, y2,
+	public GGradientPaintD(double x1, double y1, GColor color1, double x2,
+			double y2, GColor color2) {
+		impl = new GradientPaint((float) x1, (float) y1,
+				GColorD.getAwtColor(color1), (float) x2, (float) y2,
 				GColorD.getAwtColor(color2));
 	}
